[improve] add auto redirect options (#150)
Add redirection parameters. After opening, there is no need to obtain the be list, and streamloading is performed through fe.

+++ spark-doris-connector/src/main/scala/org/apache/doris/spark/writer/DorisWriter.scala
@@ -82,10 +82,10 @@ class DorisWriter(settings: SparkSettings) extends Serializable {
     doWrite(dataFrame, dorisStreamLoader.loadStream)
   }
 
-  private def doWrite(dataFrame: DataFrame, loadFunc: (util.Iterator[InternalRow], StructType) => Int): Unit = {
+  private def doWrite(dataFrame: DataFrame, loadFunc: (util.Iterator[InternalRow], StructType) => Long): Unit = {
 
     val sc = dataFrame.sqlContext.sparkContext
-    val preCommittedTxnAcc = sc.collectionAccumulator[Int]("preCommittedTxnAcc")
+    val preCommittedTxnAcc = sc.collectionAccumulator[Long]("preCommittedTxnAcc")
     if (enable2PC) {
       sc.addSparkListener(new DorisTransactionListener(preCommittedTxnAcc, dorisStreamLoader, sinkTxnIntervalMs, sinkTxnRetries))
     }
@@ -99,7 +99,7 @@ class DorisWriter(settings: SparkSettings) extends Serializable {
 
       while (iterator.hasNext) {
         val batchIterator = new BatchIterator[InternalRow](iterator, batchSize, maxRetryTimes > 0)
-        val retry = Utils.retry[Int, Exception](maxRetryTimes, Duration.ofMillis(batchInterValMs.toLong), logger) _
+        val retry = Utils.retry[Long, Exception](maxRetryTimes, Duration.ofMillis(batchInterValMs.toLong), logger) _
         retry(loadFunc(batchIterator.asJava, schema))(batchIterator.reset()) match {
           case Success(txnId) =>
             if (enable2PC) handleLoadSuccess(txnId, preCommittedTxnAcc)
@@ -116,11 +116,11 @@ class DorisWriter(settings: SparkSettings) extends Serializable {
 
   }
 
-  private def handleLoadSuccess(txnId: Int, acc: CollectionAccumulator[Int]): Unit = {
+  private def handleLoadSuccess(txnId: Long, acc: CollectionAccumulator[Long]): Unit = {
     acc.add(txnId)
   }
 
-  private def handleLoadFailure(acc: CollectionAccumulator[Int]): Unit = {
+  private def handleLoadFailure(acc: CollectionAccumulator[Long]): Unit = {
     // if task run failed, acc value will not be returned to driver,
     // should abort all pre committed transactions inside the task
     logger.info("load task failed, start aborting previously pre-committed transactions")
@@ -128,7 +128,7 @@ class DorisWriter(settings: SparkSettings) extends Serializable {
       logger.info("no pre-committed transactions, skip abort")
       return
     }
-    val abortFailedTxnIds = mutable.Buffer[Int]()
+    val abortFailedTxnIds = mutable.Buffer[Long]()
     acc.value.asScala.foreach(txnId => {
       Utils.retry[Unit, Exception](3, Duration.ofSeconds(1), logger) {
         dorisStreamLoader.abortById(txnId)
